Fix dirty ancillas for merge and slide moves (#194)
* Clean up path ancilla after slide moves

The path ancilla (needed when there are 2 or more potentially blocking
squares) could mess up the coherence of the board states, as
demonstrated by the test. This is prevented by zeroing it after the
slide is completed.

Partially fixes issue #193.

* Correct comment in quantum_moves.split_slide

* Exclude source/target qubits from split/merge path

Squares that are one of the move squares are not supposed to be included
in the path qubits for split slide or merge slide.

* Merge slide: should use ISWAP^-1 not ISWAP

* Make CirqBoard.create_path_ancilla private

* Zero ancillas for merge slide/split slide

Fixes part of issue #193.