Fix deterministically flaky tests (#201)
* puzzles_test - fix unused board parameters

* Use random but reproducible test seeds

Currently the noisy simulator uses a random generator which is
initialized to a fixed seed, but shared across tests. This means that if
you run the same test command twice in a row you'll get the same
results, but adding or modifying any test can affect the result of any
other test. An annoying kind of deterministic flakiness. (The non-noisy
simulators are currently using different seeds each time.)

This commit makes all simulators use randomly chosen seeds by default.
The seed can be forced to a specific value using the
RECIRQ_CHESS_TEST_SEED environment variable. Seeds are written to the
standard output so that when a test fails in pytest, you can see the
seed and use it to reproduce a failure.

* Fix flaky test_split_capture_with_successful_measurement_outcome

Increasing the sample size fixes it because the incorrect board is then
less likely to go over the noise threshold.

* Use noise mitigation when performing a measurement

The sample size of 1 used when getting a measurement result meant that
noise mitigation had no effect, and you could easily get an impossible
result.

This fixes the flakiness of test_a10 in puzzles_test.py.