Interactive board - don't log about start position
The interactive board prints the debug logs from the QuantumBoard object
after each move (amount of time used to perform measurements and so on).
But the board is also sampled before the first move to display the
initial position. After the first move is played, the debug logs
concerning measurement both of the position after 0 moves and the
position after 1 move.

This commit clears the logs from measuring the initial position so that
they are never displayed.