Fix duplicate allocation for singleton qubits
For a qubit that has multiple one-qubit gates and no two-qubit gates,
the layout code would map multiple qubits. The commit fixes this.

Also fix an incorrect warning and change it to an assertion.

For a 20-game test set this increased the number of half-moves able to
be mapped onto the Sycamore23 device from 461 to 470.