Fix two chess movement bugs (#202)
* Fix bug with PAWN_CAPTURE implementation

A test reproducing the bug is included.

* Remove bogus unhook for split jump

The split jump source square is not guaranteed to end empty. A test is
included which demonstrates this.

Partially fixes issue #153.

Some tests have to be disabled due to the circuit becoming larger.

* Optimization for split jump basic case

If we know that the split jump squares are in the <100| state then we
know that the source square will be empty afterward.

Re-enable the tests disabled in the previous commit.