Make quantum board more automation-friendly (#206)
* sample_with_ancilla - configure # of raw samples

The existing logic for determining how many raw samples to use is moved
to a suggest_num_reps function. The caller can specify num_reps to
sample_with_ancilla if they want to use a different value.

* Remove circuit diagram from board log

For some game with about 40 entangled squares, it was taking over 1
second to render and having over 100k characters.