Quantum chess bug fixes (#207)
* Fix Move.__eq__ (check source2)

* Interactive board - don't log about start position

The interactive board prints the debug logs from the QuantumBoard object
after each move (amount of time used to perform measurements and so on).
But the board is also sampled before the first move to display the
initial position. After the first move is played, the debug logs
concerning measurement both of the position after 0 moves and the
position after 1 move.

This commit clears the logs from measuring the initial position so that
they are never displayed.

* Fix duplicate allocation for singleton qubits

For a qubit that has multiple one-qubit gates and no two-qubit gates,
the layout code would map multiple qubits. The commit fixes this.

Also fix an incorrect warning and change it to an assertion.

For a 20-game test set this increased the number of half-moves able to
be mapped onto the Sycamore23 device from 461 to 470.