Remove 'use_cache' args to QuantumBoard functions
We discussed making caching enable/disable a class-level property. But
pre-cached moves can only be used if you manually request them to be
calculated. The feature won't have any effect if you don't intentionally
use it, so I don't think there needs to be an on/off switch.

Besides the pre-cached moves there is another kind of 'caching', which
is storing the square probabilities after a sample has been taken, and
returning the same ones if probabilities are requested again. It was
never possible to turn this off. This commit also fixes some
bugs/inconsistencies with that:
- Probabilities may be reused if sample size was *larger* than requested
- Probabilities always saved if calculated (instead of on calling the
get probabilities function)

Other refactoring changes:
- Use immutable probability vectors
- Also store full/empty squares in move_history_probabilities_cache
  (useful for cache moves which need the previous move's probabilities)
- Make entry for starting position in move_history_probabilities_cache
  (gets rid of edge case)