Run notebooks through nbfmt (#371)
Builds are failing on GitHub because some of the notebooks need to be
reformatted. Possibly this is a consequence of a version dependency
having changed somewhere. (The requirements.txt file does not pin
versions, and since the CI workflows use the same requirements.txt, a
new version of something there could influence the results.)