Use specific versions of GitHub runners and actions (#370)
Google's terms for allowing the use of GitHub Actions on Google-owned repositories requires that third-party actions be referenced using a specific commit, not a tagged release or a branch name. They also recommend that GitHub-hosted runners be referenced by fixed versions and not "-latest". (Internal doc link: go/github-actions#actions)

The SHAs for GitHub Actions in this commit were obtained using frizbee. The runner versions equivalent to the "-latest" runners are based on the table at https://github.com/actions/runner-images