Make recirq/kpz/experiment_test.py deterministic (#391)
Add constant seed so that the sampler and RNG generator are deterministic.
This will fix flakiness in test_size_independence_muinf within recirq/kpz/experiment_test.py

Note that the test can still produce different results if multi-threaded pytest is used depending on the order of execution of different tests, but this will vastly reduce the number of cases tested.