Guard against unavailable PyTket during Cirq "next" tests
PyTket pins to a specific cirq version and doesn't work with Cirq "next"
during testing. (See `dev_tools/write-ci-requirements.py`.) However, the
placement code in placement.py did not check if pytket was loaded when
RouteCQCC is not implemented, and this led to a runtime error during CI
tests.

The changes here make it check that Pytket was loaded so that it doesn't
dereference unassigned variables at run time, and also skips the
placement test when the environment variable `RECIRQ_IMPORT_FAILSAFE` is set.