Update Scorecard workflow & rename file
This updates the Scorecard workflow to run it on pull requests (instead
of only weekly). It also adds a step to print a link on the job summary
page to make it easier to look up the results. Finally, it renames the
file scorecard-scanner.yaml in an effort to be a little more clear.