Update workflows to use only 'main' branch
There's at least one security scanning workflow (Scorecard) that seems
to have an odd permissions problem if a nonexistent branch is included
in the event trigger. Now that we switch to 'main' instead of 'master',
we may as well remove mention of 'master' anyway, and hopefully that'll
also solve the weird branch warning I'm seeing from Scorecard.