Add code for Contextuality experiment (#446)
PR to include the code for the experiments in this paper: https://arxiv.org/abs/2512.02284

Functions and Classes for the 2D GHZ Game Experiment.
This module provides a complete pipeline for benchmarking the fidelity of
Greenberger-Horne-Zeilinger (GHZ) states on a 2D grid of qubits.
This is the code used for the implementation of the GHZ game detailed in arxiv.2512.02284